コンテンツ一覧機能の更新機能を改善。

+++ src/main/java/com/example/repository/ContentRepository.java
@@ -19,7 +19,7 @@ import com.example.domain.MstItem;
 public interface ContentRepository extends JpaRepository<Content, Long> {
   List<Content> findByContentName(String contentName);
   
-  Content findByContentId(int contentId);
+  Content findByContentId(Long contentId);
 
   Page<Content> findAllOrderByContentName(Pageable pageable);
