+++ src/main/resources/db/migration/V3__suppliertable.sql
@@ -32,10 +32,10 @@ INSERT INTO mst_item (item_id, item_name,item_type_id,item_expire_date,price,ver
 
 
 CREATE TABLE content (
- content_id INT NOT NULL PRIMARY KEY,
+ content_id bigint NOT NULL PRIMARY KEY,
  item_id INT,
  content_name VARCHAR(100),count INT,comment VARCHAR(500),
- version INT,
+ version bigint,
  updated_function VARCHAR(100),
  created_function VARCHAR(100),
  delete_flag BIT(1)
